Refactor Footer components
