+++ views/BoxNavigation.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { View, StyleSheet } from 'react-native'
+import { View, StyleSheet, Image } from 'react-native'
 import ContactUsButton from '../components/ContactUsButton'
 
 function BoxNavigation() {
